chore: load dialogs dynamically
