+++ apps/dashboard/app/(app)/apis/[apiId]/keys/[keyAuthId]/_components/components/table/keys-list.tsx
@@ -2,7 +2,7 @@
 import { VirtualTable } from "@/components/virtual-table/index";
 import type { Column } from "@/components/virtual-table/types";
 import type { KeyDetails } from "@/lib/trpc/routers/api/keys/query-api-keys/schema";
-import { BookBookmark, Focus, Key } from "@unkey/icons";
+import { BookBookmark, Dots, Focus, Key } from "@unkey/icons";
 import {
   AnimatedLoadingSpinner,
   Button,
@@ -14,9 +14,9 @@ import {
   TooltipTrigger,
 } from "@unkey/ui";
 import { cn } from "@unkey/ui/src/lib/utils";
+import dynamic from "next/dynamic";
 import Link from "next/link";
 import React, { useCallback, useMemo, useState } from "react";
-import { KeysTableActionPopover } from "./components/actions/keys-table-action.popover";
 import { getKeysTableActionItems } from "./components/actions/keys-table-action.popover.constants";
 import { VerificationBarChart } from "./components/bar-chart";
 import { HiddenValueCell } from "./components/hidden-value";
@@ -33,6 +33,27 @@ import {
 import { StatusDisplay } from "./components/status-cell";
 import { useKeysListQuery } from "./hooks/use-keys-list-query";
 import { getRowClassName } from "./utils/get-row-class";
+
+const KeysTableActionPopover = dynamic(
+  () =>
+    import("./components/actions/keys-table-action.popover").then((mod) => ({
+      default: mod.KeysTableActionPopover,
+    })),
+  {
+    ssr: false,
+    loading: () => (
+      <button
+        type="button"
+        className={cn(
+          "group-data-[state=open]:bg-gray-6 group-hover:bg-gray-6 group size-5 p-0 rounded m-0 items-center flex justify-center",
+          "border border-gray-6 group-hover:border-gray-8 ring-2 ring-transparent focus-visible:ring-gray-7 focus-visible:border-gray-7",
+        )}
+      >
+        <Dots className="group-hover:text-gray-12 text-gray-11" size="sm-regular" />
+      </button>
+    ),
+  },
+);
 
 export const KeysList = ({
   keyspaceId,
